fixed several issues (#328)
* fix some warnings

* no need to cast them it seems

going off of https://github.com/tmp64/BugfixedHL-Rebased

* fix invalid character somehow

idk what happened how how i fixed it

* fix invalid char

* match the cpp function def

* include APIProxy from hl-updated sdk

* fix warning for M_PI redef

* use std::size instead of ARRAYSIZE macro

* comment out ARRAYSIZE macro

* formatting

* fix another warning

* fix snprintf return issues for scripts

* fix POSX method warnings

* make AS_ChatLog use chat logger

* fix chat log file creation

* fix character select "At <map>" text

* script dbg now works again and enable debug log by default

* undo "Strip edicts on env_model and env_sprite with no targetname"

* get rid of legacy MSErrorConsoleText

* forgot to commit gg

* get rid of other legacy logging functions

+++ src/public/engine/eiface.h
@@ -192,7 +192,7 @@ typedef struct enginefuncs_s
 	void        (*pfnCRC32_ProcessBuffer)   (CRC32_t *pulCRC, void *p, int len);
 	void		(*pfnCRC32_ProcessByte)     (CRC32_t *pulCRC, unsigned char ch);
 	CRC32_t		(*pfnCRC32_Final)			(CRC32_t pulCRC);
-	long		(*pfnRandomLong)			(long  lLow,  long  lHigh);
+	int32		(*pfnRandomLong)			(int32 lLow,  int32  lHigh);
 	float		(*pfnRandomFloat)			(float flLow, float flHigh);
 	void		(*pfnSetView)				(const edict_t *pClient, const edict_t *pViewent );
 	float		(*pfnTime)					( void );
